Rename Slay.Prim to Slay.Cairo.Prim
