+++ cairo/slay-cairo.cabal
@@ -14,6 +14,7 @@ library
 
   exposed-modules:
     Slay.Cairo
+    Slay.Cairo.Prim
 
   build-depends:
     base,
